[FIRRTL] Add circuit linking infrastructure
Circuit linking allows a module in one circuit to make XMRs to a module in a
separately compiled circuit.

This is accomplished by, for the source circuit, generating an ExportableRefOp
in the LowerXMR pass for any public output probe port. In the ExportVerilog
pass, every ExportableRefOp is used to create an ExportRefOp, which is written
to an exports.mlir file.

In the client circuit, linking is enabled by supplying the firtool option
-link-circuit-exports with the path to the source circuit exports.mlir file.
In the LinkCircuit pass, any ExtModule found in the client circuit that can
be matched with the source circuit exports gets its ref port internal paths
filled out.